+++ src/components/Main/shared-components/Swiper.tsx
